backport fix to #2720 error compiling in VC9
git-svn-id: https://svn.osgeo.org/gdal/branches/1.6@15899 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/georaster/georaster_rasterband.cpp
@@ -56,7 +56,7 @@ GeoRasterRasterBand::GeoRasterRasterBand( GeoRasterDataset *poGDS,
 
     if( nLevel )
     {
-        double dfScale  = pow( 2, nLevel );
+        double dfScale  = pow( (double) 2.0, (double) nLevel );
         nRasterXSize    = (int) ceil( nRasterXSize / dfScale );
         nRasterYSize    = (int) ceil( nRasterYSize / dfScale );
     }
